Funciona, hay que refactorizar, se cambiaron los nombres de las clases de atributos para seguir la norma

+++ trifenix.agro.db.applicationsReference/agro/Common/Counters.cs
@@ -11,14 +11,14 @@ namespace trifenix.agro.db.applicationsReference.agro.Common {
 
         public async Task<long> GetCounter<T>(string query) where T : DocumentBase => await Client<T>().QuerySingleAsync<long>(query);
 
-        public async Task<long> GetLastCounterDoses(string idProduct) {
+        public async Task<int> GetLastCounterDoses(string idProduct) {
             var existAny = await SingleQuery<Dose, long>(Queries(DbQuery.COUNT_DOSES_BY_PRODUCTID), idProduct);
             if (existAny == 0)
                 return 0;
-            return await SingleQuery<Dose, long>(Queries(DbQuery.MAXCORRELATIVE_DOSES_BY_PRODUCTID), idProduct);
+            return await SingleQuery<Dose, int>(Queries(DbQuery.MAXCORRELATIVE_DOSES_BY_PRODUCTID), idProduct);
         }
 
-        public async Task<long> GetCorrelativeFromDoses(string idDoses) => await SingleQuery<Dose, long>(Queries(DbQuery.CORRELATIVE_FROM_DOSESID), idDoses);
+        public async Task<int> GetCorrelativeFromDoses(string idDoses) => await SingleQuery<Dose, int>(Queries(DbQuery.CORRELATIVE_FROM_DOSESID), idDoses);
 
     }
